Wizard component
This commit introduces a wizard component (basically an opinionated
Backbone.View) that make simplier the wizard creations. It defines steps
to build the wizard screens and creates a progress breadcrumb. It also brings
defaults styles for the wizards.

This commit also comes with a rebanded tutorial view (now called _install
wizard_) that takes care of the wizard component and defines the routes related
to it. It also remove almost (I expect all) the code related to the deprecated
_tutorial_.